Move language files to language folder
